FIX: Fix some bugs(hash password, owner's video access authority

+++ wetube/src/controllers/videoController.js
@@ -1,4 +1,5 @@
 import Video from '../models/Video';
+import User from '../models/User';
 
 export const homepage = async (req, res) => {
   try {
@@ -12,7 +13,6 @@ export const homepage = async (req, res) => {
 export const watch = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id).populate('owner');
-  console.log(video);
   if (!video) {
     return res.status(400).render('404', { pageTitle: 'Video not found' });
   }
@@ -23,10 +23,20 @@ export const watch = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
   const video = await Video.findById(id);
+  // Error if video is not found
   if (!video) {
     return res.status(404).render('404', { pageTitle: 'Video not found' });
+  }
+  // Error if video owner and user logging in are not matched
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   return res.render('videos/edit', {
     pageTitle: `Edit: ${video.title}`,
@@ -35,11 +45,21 @@ export const getEdit = async (req, res) => {
 };
 
 export const postEdit = async (req, res) => {
-  const { id } = req.params;
-  const { title, description, hashtags } = req.body;
-  const video = await Video.exists({ _id: id });
+  const {
+    params: { id },
+    body: { title, description, hashtags },
+    session: {
+      user: { _id },
+    },
+  } = req;
+  const video = await Video.findById(id);
+  // Error if video is not found
   if (!video) {
     return res.render('404', { pageTitle: 'Video not found' });
+  }
+  // Error if video owner and user logging in are not matched
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
   }
   await Video.findByIdAndUpdate(id, {
     title,
@@ -50,8 +70,25 @@ export const postEdit = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    params: { id },
+    session: {
+      user: { _id },
+    },
+  } = req;
+  const video = await Video.findById(id);
+  const user = await User.findById(_id);
+  // Error if video is not found
+  if (!video) {
+    return res.render('404', { pageTitle: 'Video not found' });
+  }
+  // Error if video owner and user logging in are not matched
+  if (String(video.owner) !== String(_id)) {
+    return res.status(403).redirect('/');
+  }
   await Video.findByIdAndDelete(id);
+  user.videos.splice(user.videos.indexOf(id), 1);
+  user.save();
   return res.redirect('/');
 };
 
@@ -62,19 +99,25 @@ export const getUpload = (req, res) => {
 export const postUpload = async (req, res) => {
   const {
     session: {
-      user: { _id },
+      user: { _id }, // user who is logging in now
     },
     file: { path: fileUrl }, // ES6; Create constants named {path} from req.file.path and change its name to fileUrl
     body: { title, description, hashtags },
   } = req;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
       description,
       fileUrl,
       owner: _id,
       hashtags: Video.formatHashtags(hashtags),
     });
+    // await User.findOneAndUpdate(_id, {
+    //   $push: {videos: {$each: newVideo._id}}
+    // });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect('/');
   } catch (error) {
     console.log(error);
